Add q-note shortcode and notes.yml files and support
So that the publication can support endbnotes, rather than footnotes, we'll be inserting the note reference numbers with a shortcode instead of markdown, and then keeping the notes themselves in a yaml file. In the interface, clicking on a number pop-ups up a window with the note text. A separate endnotes page then automatically creates an ordered list of the notes in the yaml doc.