Update config for Goldmark renderer
when we moved from v17.0 to v18.0, the markdown renderer in Hugo changed from Blackfriday to Goldmark. Goldmark requires some extra info in config.yml to allow inline HTML to be rendered. And it’s the inline HTML -- <span class="cite-current-date">Aug. 29, 2019</span> -- that makes the date appear correctly as today’s date.